TfRecord Support for Multiple Icons
Slightly modification of feature description and image parsing logic to
handle images with multiple instances of an icon (aka, multiple gold
bounding boxes.)

+++ modules/bounding_box.py
@@ -9,16 +9,16 @@ class BoundingBox:
    Assumes each coordinate corresponds to a pixel, and
     therefore is zero-indexed.
   """
-  min_x: float
-  min_y: float
-  max_x: float
-  max_y: float
+  min_x: int
+  min_y: int
+  max_x: int
+  max_y: int
 
   def __init__(self, min_x: float, min_y: float, max_x: float, max_y: float):
-    self.min_x = min_x
-    self.min_y = min_y
-    self.max_x = max_x
-    self.max_y = max_y
+    self.min_x = int(min_x)
+    self.min_y = int(min_y)
+    self.max_x = int(max_x)
+    self.max_y = int(max_y)
 
   def calculate_area(self) -> float:
     # add one in calculations because pixel numbers are 0-indexed
